Moved openAI Samples
