+++ generative_ai/openai/chat_openai_image_example.py
