+++ generative_ai/openai/chat_openai_image_stream_example.py
