+++ generative_ai/openai/chat_openai_stream_example.py
